feat: Enhance skip functionality, add defenses.

+++ skip.php
@@ -8,6 +8,7 @@
 	$SPOTIFY_CLIENT_ID = getenv('SPOTIFY_CLIENT_ID');
 	$SPOTIFY_CLIENT_SECRET = getenv('SPOTIFY_CLIENT_SECRET');
 	$SPOTIFY_REDIRECT_URI = getenv('SPOTIFY_REDIRECT_URI');
+	$SKIP_PASSWORD = getenv('SKIP_PASSWORD');
 	$SALT = getenv('SALT');
 
 	if (empty($SALT)) {
@@ -65,23 +66,33 @@
 			$api->setAccessToken($atoken);
 		}
 
+		$skipPassword = $_GET['p'];
+
 		try {
-			$api->next(0);
-			print 'skipping...';
-		} catch (Exception $e) {
-			try {
-				$session->refreshAccessToken($rtoken);
-				$accessToken = $session->getAccessToken();
-				$api->setAccessToken($accessToken);
-
-				unlink('./' . $SALT . 'access.txt');
-				$writeAccess = @file_put_contents('./' . $SALT . 'access.txt', $accessToken);
-
+			if (empty($SKIP_PASSWORD) || empty($skipPassword) || ($SKIP_PASSWORD !== $skipPassword)) {
+				throw new Exception('password');
+			} else {
 				$api->next(0);
 				print 'skipping...';
-			} catch (Exception $e) {
-				print_r($e);
-				echo 'Error: Bad session token or client (1)';
+			}
+		} catch (Exception $e) {
+			if ($e->getMessage() === 'password') {
+				echo "I'm sorry, I can't do that. (BAD PASSWORD)";
+			} else {
+				try {
+					$session->refreshAccessToken($rtoken);
+					$accessToken = $session->getAccessToken();
+					$api->setAccessToken($accessToken);
+
+					unlink('./' . $SALT . 'access.txt');
+					$writeAccess = @file_put_contents('./' . $SALT . 'access.txt', $accessToken);
+
+					$api->next(0);
+					print 'skipping...';
+				} catch (Exception $e) {
+					print_r($e);
+					echo 'Error: Bad session token or client (1)';
+				}
 			}
 		}
